Re-org auth middleware
