+++ api/server.js
@@ -2,7 +2,7 @@ const express = require('express');
 const cors = require('cors');
 const helmet = require('helmet');
 const morgan = require('morgan');
-const { authMiddleware } = require('../auth/authentication');
+const authenticator = require('../middleware/authenticator');
 
 const server = express();
 
@@ -20,6 +20,6 @@ const userRouter = require('./user');
 
 server.use('/api/auth', authRouter);
 server.use('/api/stories', storiesRouter);
-server.use('/api/user', authMiddleware, userRouter);
+server.use('/api/user', authenticator, userRouter);
 
 module.exports = server;
